Rename library whimsy
