+++ tests/main.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python2
 '''
-The main source for whimsy. Ties together the default test runners and
+The main source for testlib. Ties together the default test runners and
 loaders.
 
 Discovers and runs all tests from a given root directory.
@@ -16,9 +16,9 @@ ext_path = os.path.join(base_dir, os.pardir, 'ext')
 sys.path.insert(0, base_dir)
 sys.path.insert(0, ext_path)
 
-import whimsy.main as whimsy
-import whimsy.config as config
-import whimsy.helper as helper
+import testlib.main as testlib
+import testlib.config as config
+import testlib.helper as helper
 
 config.basedir = helper.absdirpath(__file__)
-whimsy()
+testlib()
